Develop to master

See merge request techspecialist/procureease-api!44

+++ ProcureEaseAPI/Controllers/UsersController.cs
@@ -194,20 +194,15 @@ namespace ProcureEaseAPI.Controllers
                 {
                     success = true,
                     message = "User added successfully.",
-                    data = db.UserProfile.Where(x => x.TenantID == tenantId).Select(x => new
+                    data = db.UserProfile.Where(x => x.Catalog.TenantID == tenantId).Select(x => new
                     {
-                        User = new
-                        {
-                            x.UserID,
-                            FullName = x.FirstName + " " + x.LastName
-                        },
-                        Department = new
-                        {
-                            x.DepartmentID,
-                            x.Department1.DepartmentName
-                        }
-
-                    }),
+                        x.UserID,
+                        FullName = x.FirstName + " " + x.LastName,
+                        x.Department1.DepartmentName,
+                        x.DepartmentID,
+                        x.UserEmail,
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                    })
                 }, JsonRequestBehavior.AllowGet);
             }
             catch (Exception ex)
@@ -269,18 +264,13 @@ namespace ProcureEaseAPI.Controllers
                     message = "Please check your email to reset password.",
                     data = db.UserProfile.Where(x => x.Catalog.SubDomain == SubDomain).Select(x => new
                     {
-                        User = new
-                        {
-                            x.UserID,
-                            FullName = x.FirstName + " " + x.LastName
-                        },
-                        Department = new
-                        {
-                            x.DepartmentID,
-                            x.Department1.DepartmentName
-                        }
-
-                    }),
+                        x.UserID,
+                        FullName = x.FirstName + " " + x.LastName,
+                        x.Department1.DepartmentName,
+                        x.DepartmentID,
+                        x.UserEmail,
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                    })
                 }, JsonRequestBehavior.AllowGet);
             }
             catch(Exception ex)
@@ -330,20 +320,15 @@ namespace ProcureEaseAPI.Controllers
                 {
                     success = true,
                     message = "Password reset successful.",
-                    data = db.UserProfile.Where(x => x.Catalog.SubDomain == SubDomain).Select(x => new
+                    data = db.UserProfile.Where(x =>x.Catalog.SubDomain == SubDomain).Select(x => new
                     {
-                        User = new
-                        {
-                            x.UserID,
-                            FullName = x.FirstName + " " + x.LastName
-                        },
-                        Department = new
-                        {
-                            x.DepartmentID,
-                            x.Department1.DepartmentName
-                        }
-
-                    }),
+                        x.UserID,
+                        FullName = x.FirstName + " " + x.LastName,
+                        x.Department1.DepartmentName,
+                        x.DepartmentID,
+                        x.UserEmail,
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                    })
                 }, JsonRequestBehavior.AllowGet);
             }
             catch (Exception ex)
@@ -425,18 +410,13 @@ namespace ProcureEaseAPI.Controllers
                     message = "Sign up successful.",
                     data = db.UserProfile.Where(x => x.Catalog.SubDomain == SubDomain).Select(x => new
                     {
-                        User = new
-                        {
-                            x.UserID,
-                            FullName = x.FirstName + " " + x.LastName
-                        },
-                        Department = new
-                        {
-                            x.DepartmentID,
-                            x.Department1.DepartmentName
-                        }
-
-                    }),
+                        x.UserID,
+                        FullName = x.FirstName + " " + x.LastName,
+                        x.Department1.DepartmentName,
+                        x.DepartmentID,
+                        x.UserEmail,
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                    })
                 }, JsonRequestBehavior.AllowGet);
             }
             catch(Exception ex)
@@ -540,8 +520,7 @@ namespace ProcureEaseAPI.Controllers
                     }
                     if (findDepartment.DepartmentName != "Procurement" && findDepartment.DepartmentHeadUserID != userProfile.UserID)
                     {
-                        EditToEmployeeRole(userProfile, Id);
-                        return Json("good", JsonRequestBehavior.AllowGet);
+                        EditToEmployeeRole(userProfile, Id);                    
                     }
                     if (findDepartment.DepartmentName != "Procurement" && findDepartment.DepartmentHeadUserID == userProfile.UserID)
                     {
@@ -549,6 +528,7 @@ namespace ProcureEaseAPI.Controllers
                     }
                     Response.StatusCode = (int)HttpStatusCode.OK;                  
                 }
+                Response.StatusCode = (int)HttpStatusCode.OK;
                 return Json(new
                 {
                     success = true,
@@ -560,7 +540,7 @@ namespace ProcureEaseAPI.Controllers
                         x.Department1.DepartmentName,
                         x.DepartmentID,
                         x.UserEmail,
-                        DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                     })
                 }, JsonRequestBehavior.AllowGet);
             }
@@ -585,16 +565,32 @@ namespace ProcureEaseAPI.Controllers
             UserProfile EditProfile = db.UserProfile.Where(x => x.UserID == UserProfile.UserID).FirstOrDefault();
             EditProfile.UserEmail = UserProfile.UserEmail;
             EditProfile.DepartmentID = UserProfile.DepartmentID;
-            AspNetUserRoles role = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
-            db.AspNetUserRoles.Remove(role);
-            db.SaveChanges();
-            AspNetUserRoles userRole = new AspNetUserRoles();
-            userRole.UserId = Id;
-            userRole.RoleId = RoleId;
-            db.AspNetUserRoles.Add(userRole);
-            AspNetUsers EditUser = db.AspNetUsers.Where(x => x.Id == Id).FirstOrDefault();
-            EditUser.UserName = UserProfile.UserEmail;
-            db.SaveChanges();
+            var getUserRoleId = db.AspNetUserRoles.Where(x => x.UserId == Id).Select(x => x.RoleId);
+            AspNetUserRoles role = db.AspNetUserRoles.Find(Id, getUserRoleId);
+            if (role == null)
+            {
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                db.SaveChanges();
+            }
+            else
+            {
+                AspNetUserRoles roles = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
+                db.AspNetUserRoles.Remove(role);
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                db.SaveChanges();
+            }
         }
 
         protected void EditToHeadOfProcumentRole(UserProfile UserProfile, string Id)
@@ -603,16 +599,32 @@ namespace ProcureEaseAPI.Controllers
             UserProfile EditProfile = db.UserProfile.Where(x => x.UserID == UserProfile.UserID).FirstOrDefault();
             EditProfile.UserEmail = UserProfile.UserEmail;
             EditProfile.DepartmentID = UserProfile.DepartmentID;
-            AspNetUserRoles role = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
-            db.AspNetUserRoles.Remove(role);
-            db.SaveChanges();
-            AspNetUserRoles userRole = new AspNetUserRoles();
-            userRole.UserId = Id;
-            userRole.RoleId = RoleId;
-            db.AspNetUserRoles.Add(userRole);
-            AspNetUsers EditUser = db.AspNetUsers.Where(x => x.Id == Id).FirstOrDefault();
-            EditUser.UserName = UserProfile.UserEmail;            
-            db.SaveChanges();
+            var getUserRoleId = db.AspNetUserRoles.Where(x => x.UserId == Id).Select(x => x.RoleId);
+            AspNetUserRoles role = db.AspNetUserRoles.Find(Id, getUserRoleId);
+            if (role == null)
+            {
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                db.SaveChanges();
+            }
+            else
+            {
+                AspNetUserRoles roles = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
+                db.AspNetUserRoles.Remove(role);
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                db.SaveChanges();
+            }
         }
 
         protected void EditToEmployeeRole(UserProfile UserProfile, string Id)
@@ -621,17 +633,34 @@ namespace ProcureEaseAPI.Controllers
             UserProfile EditProfile = db.UserProfile.SingleOrDefault(x => x.UserID == UserProfile.UserID);
             EditProfile.UserEmail = UserProfile.UserEmail;
             EditProfile.DepartmentID = UserProfile.DepartmentID;
-            //AspNetUserRoles role = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
-            //db.AspNetUserRoles.Remove(role);
-            //db.SaveChanges();
-            AspNetUserRoles userRole = new AspNetUserRoles();
-            userRole.UserId = Id;
-            userRole.RoleId = RoleId;
-            db.AspNetUserRoles.Add(userRole);
-            AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
-            EditUser.UserName = UserProfile.UserEmail;
-            EditUser.Email = UserProfile.UserEmail;
+            var getUserRoleId = db.AspNetUserRoles.Where(x => x.UserId == Id).Select(x => x.RoleId);
+            AspNetUserRoles role = db.AspNetUserRoles.Find(Id,getUserRoleId);
+            if(role==null)
+            {
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                db.SaveChanges();
+            }
+            else
+            {
+                AspNetUserRoles roles = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
+                db.AspNetUserRoles.Remove(role);
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                db.SaveChanges();
+            }
             db.SaveChanges();
+        
         }
 
         protected void EditToHeadOfDepartmentRole(UserProfile UserProfile, string Id)
@@ -640,15 +669,32 @@ namespace ProcureEaseAPI.Controllers
             UserProfile EditProfile = db.UserProfile.Where(x => x.UserID == UserProfile.UserID).FirstOrDefault();
             EditProfile.UserEmail = UserProfile.UserEmail;
             EditProfile.DepartmentID = UserProfile.DepartmentID;
-            AspNetUserRoles role = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
-            db.AspNetUserRoles.Remove(role);
-            db.SaveChanges();
-            AspNetUserRoles userRole = new AspNetUserRoles();
-            userRole.UserId = Id;
-            userRole.RoleId = RoleId;
-            AspNetUsers EditUser = db.AspNetUsers.Where(x => x.Id == Id).FirstOrDefault();
-            EditUser.UserName = UserProfile.UserEmail;
-            db.SaveChanges();
+            var getUserRoleId = db.AspNetUserRoles.Where(x => x.UserId == Id).Select(x => x.RoleId);
+            AspNetUserRoles role = db.AspNetUserRoles.Find(Id, getUserRoleId);
+            if (role == null)
+            {
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                db.SaveChanges();
+            }
+            else
+            {
+                AspNetUserRoles roles = db.AspNetUserRoles.SingleOrDefault(x => x.UserId == Id);
+                db.AspNetUserRoles.Remove(role);
+                AspNetUserRoles userRole = new AspNetUserRoles();
+                userRole.UserId = Id;
+                userRole.RoleId = RoleId;
+                db.AspNetUserRoles.Add(userRole);
+                AspNetUsers EditUser = db.AspNetUsers.SingleOrDefault(x => x.Id == Id);
+                EditUser.UserName = UserProfile.UserEmail;
+                EditUser.Email = UserProfile.UserEmail;
+                db.SaveChanges();
+            }
         }
 
         //PUT: Users/UpdateUserProfile
@@ -709,7 +755,7 @@ namespace ProcureEaseAPI.Controllers
                         x.Department1.DepartmentName,
                         x.DepartmentID,
                         x.UserEmail,
-                        DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                     })
                 }, JsonRequestBehavior.AllowGet);
             }
@@ -876,7 +922,7 @@ namespace ProcureEaseAPI.Controllers
                         x.Department1.DepartmentName,
                         x.DepartmentID,
                         x.UserEmail,
-                        DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                     })
                 }, JsonRequestBehavior.AllowGet);
             }
@@ -918,7 +964,7 @@ namespace ProcureEaseAPI.Controllers
                         x.Department1.DepartmentName,
                         x.DepartmentID,
                         x.UserEmail,
-                        DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                     })
                 }, JsonRequestBehavior.AllowGet);
             } else
@@ -949,7 +995,7 @@ namespace ProcureEaseAPI.Controllers
                         x.Department1.DepartmentName,
                         x.DepartmentID,
                         x.UserEmail,
-                        DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                        DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                     })
                 }, JsonRequestBehavior.AllowGet);
             }
@@ -1021,7 +1067,7 @@ namespace ProcureEaseAPI.Controllers
                             x.Department1.DepartmentName,
                             x.DepartmentID,
                             x.UserEmail,
-                            DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                            DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                         })
                     }, JsonRequestBehavior.AllowGet);
                 }
@@ -1054,7 +1100,7 @@ namespace ProcureEaseAPI.Controllers
                             x.Department1.DepartmentName,
                             x.DepartmentID,
                             x.UserEmail,
-                            DepartmentHeadUserID = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
+                            DepartmentHeadStatus = db.UserProfile.Where(y => y.Department1.DepartmentHeadUserID == x.UserID).Select(y => (true) || (false)).FirstOrDefault()
                         })
                     }, JsonRequestBehavior.AllowGet);
                 }
